Separated business logic
